+++ lib/apollo-federation/object.rb
@@ -32,6 +32,16 @@ module ApolloFederation
         add_directive(name: 'tag', arguments: [name: 'name', values: name])
       end
 
+      def policy(policies)
+        add_directive(
+          name: 'policy',
+          arguments: [
+            name: 'policies',
+            values: policies,
+          ],
+        )
+      end
+
       def key(fields:, camelize: true, resolvable: true)
         arguments = [
           name: 'fields',